Abstract overlay datetime component
