+++ apps/website/src/pages/stream/overlay.tsx
@@ -1,73 +1,16 @@
 import { type NextPage } from "next";
-import { useEffect, useRef, useState } from "react";
-
-import { DATETIME_ALVEUS_ZONE, formatDateTimeParts } from "@/utils/datetime";
 
 import Event from "@/components/overlay/Event";
 import Weather from "@/components/overlay/Weather";
 import Timecode from "@/components/overlay/Timecode";
+import Datetime from "@/components/overlay/Datetime";
 
 const OverlayPage: NextPage = () => {
-  // Get the current time and date
-  // Refresh every 250ms
-  const [time, setTime] = useState<{
-    time: string;
-    date: string;
-  }>();
-  const timeInterval = useRef<NodeJS.Timeout>(null);
-  useEffect(() => {
-    const updateTime = () => {
-      const date = new Date();
-      const parts = formatDateTimeParts(
-        date,
-        { style: "short", time: "seconds", timezone: true },
-        { zone: DATETIME_ALVEUS_ZONE },
-      );
-
-      // Get the hour -> second parts
-      const hourIdx = parts.findIndex((part) => part.type === "hour");
-      const secondIdx = parts.findIndex((part) => part.type === "second");
-      if (hourIdx === -1 || secondIdx === -1) return;
-      const timeParts = parts
-        .slice(hourIdx, secondIdx + 1)
-        .map((part) => part.value);
-
-      // Get the AM/PM part and the timezone
-      const dayPeriod = parts.find((part) => part.type === "dayPeriod");
-      if (dayPeriod) timeParts.push(" ", dayPeriod.value);
-      const timezone = parts.find((part) => part.type === "timeZoneName");
-      if (timezone) timeParts.push(" ", timezone.value);
-
-      // Get the date
-      const year = parts.find((part) => part.type === "year");
-      const month = parts.find((part) => part.type === "month");
-      const day = parts.find((part) => part.type === "day");
-      if (!year || !month || !day) return;
-
-      setTime({
-        time: timeParts.join(""),
-        date: [year, month, day]
-          .map((part) => part.value.padStart(2, "0"))
-          .join("-"),
-      });
-    };
-
-    updateTime();
-    timeInterval.current = setInterval(updateTime, 250);
-    return () => clearInterval(timeInterval.current ?? undefined);
-  }, []);
-
-  // This can be a client-side only page
-  if (!time) return null;
-
   return (
     <div className="h-screen w-full">
-      <div className="absolute right-2 top-2 flex flex-col gap-1 text-right font-mono font-medium text-white text-stroke-2">
-        <p className="text-4xl">{time.time}</p>
-        <p className="text-4xl">{time.date}</p>
-
+      <Datetime className="absolute right-2 top-2 text-right">
         <Weather />
-      </div>
+      </Datetime>
 
       <Event className="absolute bottom-2 left-2" />
